Search grid now performs bound-checking
Still need to plot it but it is working and performs checks on the
boundaries of the search space (if we need to re-run it for we ran it on
a very small search region). Sometimes, all it would be required would
be incresing the precision since it might have skipped a better point in
the middle of the used search space.